requirements.txt
tensorflow
numpy
tensorflow-addons
gin-config
tqdm
imageio
gdown